fixed returns computed in Double type instead of Float

+++ RL/actor_critic.py
@@ -36,7 +36,7 @@ class ActorCriticAgent(DeepAgent):
         for r in reversed(self.rewards):
             r_sum = r_sum * self.y + r
             G.append(r_sum)
-        G = torch.tensor(list(reversed(G))).to(self.device)
+        G = torch.tensor(list(reversed(G)), dtype=torch.float32).to(self.device)
         G -= G.mean()
         if len(G) > 1:
             G /= (G.std() + self.eps)
